OPENIG-9224 OPENIG-9169 Refactoring to extract ProcessRegistration into other components, migrate and align all cert validation - OPENIG-9224: Extract or remove ProcessRegistration script logic into other components and migrate to IG. Remainder exists as `AmTransformFilter` - OPENIG-9169: Migrate TransportCertValidator and related Filter classes and align functionality to ensure consistent with each other - Align with other refactoring changes - new contexts and FAPI filter chains - Remove Filter heaplets and ensure only available via new FAPI filter chains
