+++ config/7.3.0/fapi1part2adv/ig/config/dev/config/config.json
@@ -275,10 +275,6 @@
       }
     },
     {
-      "name": "TransportCertValidator",
-      "type": "DefaultTransportCertValidator"
-    },
-    {
       "name": "IdmService",
       "type": "IdmService",
       "config": {
