+++ config/7.3.0/fapi1part2adv/ig/config/prod/config/config.json
@@ -263,10 +263,6 @@
       }
     },
     {
-      "name": "TransportCertValidator",
-      "type": "DefaultTransportCertValidator"
-    },
-    {
       "name": "IdmService",
       "type": "IdmService",
       "config": {
