+++ config/7.3.0/fapi1part2adv/ig/routes/routes-service/20-as-dcr-endpoint.json
@@ -24,21 +24,6 @@
           }
         },
         {
-          "comment": "Pre process registration request",
-          "name": "ProcessRegistration",
-          "type": "ScriptableFilter",
-          "config": {
-            "type": "application/x-groovy",
-            "file": "ProcessRegistration.groovy",
-            "args": {
-              "jwkSetService": "${heap['JwkSetService']}",
-              "allowIgIssuedTestCerts": "${security.enableTestTrustedDirectory}",
-              "tokenEndpointAuthMethodsSupported": "${oauth2.tokenEndpointAuthMethodsSupported}",
-              "trustedDirectoryService": "${heap['TrustedDirectoryService']}"
-            }
-          }
-        },
-        {
           "name": "ManageApiClientFilter",
           "type": "ManageApiClientFilter",
           "config": {
