+++ config/7.3.0/fapi1part2adv/ig/routes/routes-service/21-as-token-endpoint.json
@@ -54,16 +54,6 @@
           }
         },
         {
-          "comment": "Verify that the client's transport cert is valid and is mapped to their SSA",
-          "name": "TokenEndpointTransportCertValidationFilter",
-          "type": "TokenEndpointTransportCertValidationFilter",
-          "config": {
-            "trustedDirectoryService": "TrustedDirectoryService",
-            "jwkSetService": "JwkSetService",
-            "transportCertValidator": "TransportCertValidator"
-          }
-        },
-        {
           "name": "ErrorResponseFilter",
           "type": "ScriptableFilter",
           "config": {
