+++ config/7.3.0/fapi1part2adv/ig/routes/routes-service/22-as-par-endpoint.json
@@ -35,16 +35,6 @@
           }
         },
         {
-          "name": "ParEndpointTransportCertValidationFilter",
-          "type": "ParEndpointTransportCertValidationFilter",
-          "comment": "If the client provides a certificate for authentication, verify that it is valid and is mapped to their SSA",
-          "config": {
-            "trustedDirectoryService": "TrustedDirectoryService",
-            "jwkSetService": "JwkSetService",
-            "transportCertValidator": "TransportCertValidator"
-          }
-        },
-        {
           "name": "ErrorResponseFilter",
           "type": "ScriptableFilter",
           "config": {
