+++ secure-api-gateway-ig-extensions/src/test/java/com/forgerock/sapi/gateway/dcr/filter/TokenEndpointResponseFetchApiClientFilterTest.java
@@ -38,7 +38,6 @@ import org.forgerock.openig.heap.Name;
 import org.junit.jupiter.api.Test;
 
 import com.forgerock.sapi.gateway.dcr.filter.TokenEndpointResponseFetchApiClientFilter.Heaplet;
-import com.forgerock.sapi.gateway.mtls.ResponsePathTransportCertValidationFilterTest;
 import com.forgerock.sapi.gateway.util.CryptoUtils;
 import com.nimbusds.jose.JWSAlgorithm;
 
@@ -64,7 +63,14 @@ public class TokenEndpointResponseFetchApiClientFilterTest extends BaseResponseP
 
     @Override
     protected Response createValidUpstreamResponse() {
-        return ResponsePathTransportCertValidationFilterTest.createResponseWithAccessToken(CLIENT_ID);
+        return createResponseWithAccessToken(CLIENT_ID);
+    }
+
+    public static Response createResponseWithAccessToken(String clientId) {
+        Response response = new Response(Status.OK);
+        JsonValue jsonResponse = json(object(field("access_token", createAccessToken(Map.of("aud", clientId)))));
+        response.setEntity(jsonResponse);
+        return response;
     }
 
     @Test
